1. 新建UploadWebfolderlog表，用于存储上传过的网页收藏夹文件 2. 完成导入网页收藏夹，引入AngleSharp处理浏览器导出的书签文件 3. 导入书签数据到系统已完成三分之一的工作量，算法难点基本已理清

+++ WebfolderService/Model/GroupUser.cs
@@ -33,7 +33,7 @@ namespace WebfolderService.Model
         /// <summary>
         /// 是否通过
         /// </summary>
-		public ushort IsPass{get;set;}
+		public sbyte IsPass{get;set;}
             
         /// <summary>
         /// 创建时间
